Break up Puzzle implementation into smaller files
