+++ include/ball_sort/puzzle.hpp
@@ -28,12 +28,11 @@ class Puzzle {
  private:
     void validate_puzzle() const;
 
-    [[nodiscard]] static constexpr Tubes
-    make_tubes(const Strings& letter_strings);
+    [[nodiscard]] static Tubes make_tubes(const Strings& letter_strings);
     [[nodiscard]] static Tubes make_tubes(const std::string& number_string);
 
     [[nodiscard]] std::unordered_map<char, size_t> get_ball_tally() const;
-    [[nodiscard]] constexpr std::string get_serialised_state() const;
+    [[nodiscard]] std::string get_serialised_state() const;
     [[nodiscard]] bool is_legal_move(size_t origin_index,
                                      size_t destination_index) const;
